+++ src/Action.php
@@ -3,73 +3,31 @@
 namespace Laraditz\Action;
 
 use BadMethodCallException;
-use Illuminate\Http\Request;
-use Illuminate\Support\Arr;
-use Laraditz\Action\Traits\Resolvable;
 
 class Action
 {
-    use Resolvable;
-
-    public $attributes = [];
-
-    public function __construct()
+    public function data(): array
     {
-        if (func_num_args() > 0) {
-            $this->resolveConstructorAttributes(...func_get_args());
-        }
-    }
+        $body = [];
+        $class = new \ReflectionClass(static::class);
 
-    protected function handleNow(array $attributes = [])
-    {
-        $this->fill($attributes);
-        $this->resolveRules();
+        $constructor = $class->getConstructor();
 
-        if (method_exists($this, 'handle')) {
-            return $this->resolveMethod($this, 'handle');
-        }
-    }
+        foreach ($constructor->getParameters() as $property) {
 
-    public function fill(array $attributes)
-    {
-        $this->attributes = array_merge($this->attributes, $attributes);
+            if ($property->allowsNull() === false || $this->{$property->name}) {
 
-        return $this;
-    }
-
-    public function all()
-    {
-        return $this->attributes;
-    }
-
-    public function validated()
-    {
-        return $this->validator->validated();
-    }
-
-    public function __set($key, $value)
-    {
-        Arr::set($this->attributes, $key, $value);
-    }
-
-    public function __get($key)
-    {
-        return Arr::get($this->attributes, $key, null);
-    }
-
-    public function __invoke(array $attributes = [])
-    {
-        if (app(Request::class)->all()) {
-            $attributes = array_merge($attributes, app(Request::class)->all());
+                $body += [$property->name => $this->{$property->name}];
+            }
         }
 
-        return $this->now($attributes);
+        return $body;
     }
 
     public function __call($method, $arguments)
     {
-        if ($method === 'now' || $method === 'dispatch') {
-            return $this->handleNow(...$arguments);
+        if ($method === 'run') {
+            return $this->handle(...$arguments);
         }
 
         throw new BadMethodCallException(sprintf(
@@ -79,10 +37,11 @@ class Action
         ));
     }
 
+
     public static function __callStatic($method, $arguments)
     {
-        if ($method === 'now' || $method === 'dispatch') {
-            return (new static(...$arguments))->handleNow();
+        if ($method === 'run') {
+            return (new static(...$arguments))->handle();
         }
 
         return (new static)->$method(...$arguments);